refactor(onboarding): extract TryASearch button to own class file
