+++ app/src/main/java/com/duckduckgo/app/browser/BrowserTabFragment.kt
@@ -5339,7 +5339,7 @@ class BrowserTabFragment :
 
         private fun showDaxOnboardingBubbleCta(configuration: DaxBubbleCta) {
             hideNewTab()
-            val container = if (configuration is DaxBubbleCta.BrandDesignBubbleCta) {
+            val container = if (configuration is DaxBubbleCta.BrandDesignUpdateBubbleCta) {
                 daxDialogIntroBubble.root.gone()
                 brandDesignDialogScrollView.show()
                 daxDialogIntroBubbleBrandDesign.daxCtaContainer
