+++ app/src/main/java/com/duckduckgo/app/cta/ui/Cta.kt
@@ -46,7 +46,6 @@ import com.duckduckgo.app.global.install.AppInstallStore
 import com.duckduckgo.app.global.install.daysInstalled
 import com.duckduckgo.app.onboarding.store.OnboardingStore
 import com.duckduckgo.app.onboarding.ui.view.DaxTypeAnimationTextView
-import com.duckduckgo.app.onboarding.ui.view.OnboardingSelectionButton
 import com.duckduckgo.app.pixels.AppPixelName
 import com.duckduckgo.app.pixels.AppPixelName.SITE_NOT_WORKING_SHOWN
 import com.duckduckgo.app.pixels.AppPixelName.SITE_NOT_WORKING_WEBSITE_BROKEN
@@ -791,7 +790,7 @@ sealed class DaxBubbleCta(
         appInstallStore = appInstallStore,
     )
 
-    abstract class BrandDesignBubbleCta(
+    abstract class BrandDesignUpdateBubbleCta(
         ctaId: CtaId,
         @StringRes title: Int,
         @StringRes description: Int,
@@ -1035,63 +1034,6 @@ sealed class DaxBubbleCta(
 
         override fun hideDaxBubbleCta(binding: FragmentBrowserTabBinding) {
             binding.includeNewBrowserTab.includeOnboardingDaxDialogBubbleBrandDesignUpdate.root.gone()
-        }
-    }
-
-    data class DaxIntroSearchOptionsBrandDesignUpdateCta(
-        override val onboardingStore: OnboardingStore,
-        override val appInstallStore: AppInstallStore,
-        override val isLightTheme: Boolean,
-    ) : BrandDesignBubbleCta(
-        ctaId = CtaId.DAX_INTRO,
-        title = R.string.onboardingSearchDaxDialogTitle,
-        description = R.string.onboardingSearchDaxDialogDescription,
-        options = onboardingStore.getSearchOptions(),
-        backgroundRes = R.drawable.bg_onboarding_search_options,
-        shownPixel = AppPixelName.ONBOARDING_DAX_CTA_SHOWN,
-        okPixel = AppPixelName.ONBOARDING_DAX_CTA_OK_BUTTON,
-        ctaPixelParam = Pixel.PixelValues.DAX_INITIAL_CTA,
-        onboardingStore = onboardingStore,
-        appInstallStore = appInstallStore,
-        isLightTheme = isLightTheme,
-    ) {
-        override val activeIncludeId: Int = R.id.optionsContent
-
-        override fun configureContentViews(view: View) {
-            val optionViews: List<OnboardingSelectionButton> = listOf(
-                view.findViewById(R.id.brandDesignOption1),
-                view.findViewById(R.id.brandDesignOption2),
-                view.findViewById(R.id.brandDesignOption3),
-                view.findViewById(R.id.brandDesignOption4),
-            )
-            options?.let {
-                optionViews.forEachIndexed { index, buttonView ->
-                    buttonView.show()
-                    if (it.size > index) {
-                        it[index].setOptionView(buttonView)
-                    } else {
-                        buttonView.gone()
-                    }
-                }
-            }
-        }
-
-        override fun setOnOptionClicked(
-            onboardingExperimentEnabled: Boolean,
-            configuration: DaxBubbleCta?,
-            onOptionClicked: (DaxDialogIntroOption, index: Int?) -> Unit,
-        ) {
-            options?.forEachIndexed { index, option ->
-                val optionViewId = when (index) {
-                    0 -> R.id.brandDesignOption1
-                    1 -> R.id.brandDesignOption2
-                    2 -> R.id.brandDesignOption3
-                    else -> R.id.brandDesignOption4
-                }
-                ctaView?.findViewById<MaterialButton>(optionViewId)?.setOnClickListener {
-                    onOptionClicked.invoke(option, index)
-                }
-            }
         }
     }
